+++ examples/execute_everything.py
@@ -4,6 +4,12 @@ import sys
 import traceback
 
 
+def ocel_occm_example():
+    from examples import ocel_occm_example
+    print("\n\nocel_occm_example")
+    ocel_occm_example.execute_script()
+
+
 def ocel_clustering():
     from examples import ocel_clustering
     print("\n\nocel_clustering")
@@ -14,12 +20,6 @@ def ocel_enrichment():
     from examples import ocel_enrichment
     print("\n\nocel_enrichment")
     ocel_enrichment.execute_script()
-
-
-def openai_log_queries():
-    from examples import openai_log_queries
-    print("\n\nopenai_log_queries")
-    openai_log_queries.execute_script()
 
 
 def validation_ocel20_xml():
@@ -772,9 +772,10 @@ if __name__ == "__main__":
     execute_script(inductive_miner_dfg)
     execute_script(inductive_miner_variants)
     execute_script(heu_miner_plus_plus)
+    execute_script(ocel_occm_example)
     execute_script(ocel_clustering)
     execute_script(ocel_enrichment)
-    execute_script(openai_log_queries)
+    execute_script(openai_queries)
     execute_script(validation_ocel20_xml)
     execute_script(consecutive_act_case_grouping_filter)
     execute_script(cost_based_dfg)