+++ pm4py/objects/ocel/exporter/sqlite/variants/pandas_exporter.py
@@ -17,6 +17,8 @@
 
 from pm4py.objects.ocel.obj import OCEL
 from typing import Dict, Any, Optional
+from pm4py.objects.ocel.util import ocel_consistency
+import os
 
 
 def apply(ocel: OCEL, target_path: str, parameters: Optional[Dict[Any, Any]] = None):
@@ -36,6 +38,11 @@ def apply(ocel: OCEL, target_path: str, parameters: Optional[Dict[Any, Any]] = N
         parameters = {}
 
     import sqlite3
+
+    if os.path.exists(target_path):
+        os.remove(target_path)
+
+    ocel = ocel_consistency.apply(ocel, parameters=parameters)
 
     conn = sqlite3.connect(target_path)