+++ pm4py/vis.py
@@ -49,7 +49,7 @@ def view_petri_net(petri_net: PetriNet, initial_marking: Optional[Marking] = Non
     :param petri_net: Petri net
     :param initial_marking: Initial marking
     :param final_marking: Final marking
-    :param format: Format of the output picture
+    :param format: Format of the output picture (if html is provided, GraphvizJS is used to render the visualization in an HTML page)
     :param bgcolor: Background color of the visualization (default: white)
     :param decorations: Decorations (color, label) associated to the elements of the Petri net
 
@@ -102,7 +102,7 @@ def view_performance_dfg(dfg: dict, start_activities: dict, end_activities: dict
     :param dfg: DFG object
     :param start_activities: Start activities
     :param end_activities: End activities
-    :param format: Format of the output picture
+    :param format: Format of the output picture (if html is provided, GraphvizJS is used to render the visualization in an HTML page)
     :param aggregation_measure: Aggregation measure (default: mean): mean, median, min, max, sum, stdev
     :param bgcolor: Background color of the visualization (default: white)
 
@@ -168,7 +168,7 @@ def view_dfg(dfg: dict, start_activities: dict, end_activities: dict, format: st
     :param dfg: DFG object
     :param start_activities: Start activities
     :param end_activities: End activities
-    :param format: Format of the output picture
+    :param format: Format of the output picture (if html is provided, GraphvizJS is used to render the visualization in an HTML page)
     :param bgcolor: Background color of the visualization (default: white)
 
     .. code-block:: python3
@@ -227,7 +227,7 @@ def view_process_tree(tree: ProcessTree, format: str = constants.DEFAULT_FORMAT_
     Views a process tree
 
     :param tree: Process tree
-    :param format: Format of the visualization
+    :param format: Format of the visualization (if html is provided, GraphvizJS is used to render the visualization in an HTML page)
     :param bgcolor: Background color of the visualization (default: white)
 
     .. code-block:: python3
@@ -295,7 +295,7 @@ def view_bpmn(bpmn_graph: BPMN, format: str = constants.DEFAULT_FORMAT_GVIZ_VIEW
     Views a BPMN graph
 
     :param bpmn_graph: BPMN graph
-    :param format: Format of the visualization
+    :param format: Format of the visualization (if html is provided, GraphvizJS is used to render the visualization in an HTML page)
     :param bgcolor: Background color of the visualization (default: white)
 
     .. code-block:: python3
@@ -379,7 +379,7 @@ def __dotted_attribute_selection(log: Union[EventLog, pd.DataFrame], attributes)
 
 
 @deprecation.deprecated(deprecated_in="2.3.0", removed_in="3.0.0", details="the dotted chart visualization will be removed in a future release.")
-def view_dotted_chart(log: Union[EventLog, pd.DataFrame], format: str = "png", attributes=None, bgcolor: str = "white"):
+def view_dotted_chart(log: Union[EventLog, pd.DataFrame], format: str = "png", attributes=None, bgcolor: str = "white", show_legend: bool = True):
     """
     Displays the dotted chart
 
@@ -402,6 +402,8 @@ def view_dotted_chart(log: Union[EventLog, pd.DataFrame], format: str = "png", a
     :param log: Event log
     :param format: Image format
     :param attributes: Attributes that should be used to construct the dotted chart. If None, the default dotted chart will be shown: x-axis: time y-axis: cases (in order of occurrence in the event log) color: activity. For custom attributes, use a list of attributes of the form [x-axis attribute, y-axis attribute, color attribute], e.g., ["concept:name", "org:resource", "concept:name"])
+    :param bgcolor: background color to be used in the dotted chart
+    :param show_legend: boolean (enables/disables showing the legend)
 
     .. code-block:: python3
 
@@ -418,13 +420,19 @@ def view_dotted_chart(log: Union[EventLog, pd.DataFrame], format: str = "png", a
         check_pandas_dataframe_columns(log)
 
     log, attributes = __dotted_attribute_selection(log, attributes)
+
+    parameters = {}
+    parameters["format"] = format
+    parameters["bgcolor"] = bgcolor
+    parameters["show_legend"] = show_legend
+
     from pm4py.visualization.dotted_chart import visualizer as dotted_chart_visualizer
-    gviz = dotted_chart_visualizer.apply(log, attributes, parameters={"format": format, "bgcolor": bgcolor})
+    gviz = dotted_chart_visualizer.apply(log, attributes, parameters=parameters)
     dotted_chart_visualizer.view(gviz)
 
 
 @deprecation.deprecated(deprecated_in="2.3.0", removed_in="3.0.0", details="the dotted chart visualization will be removed in a future release.")
-def save_vis_dotted_chart(log: Union[EventLog, pd.DataFrame], file_path: str, attributes=None, bgcolor: str = "white"):
+def save_vis_dotted_chart(log: Union[EventLog, pd.DataFrame], file_path: str, attributes=None, bgcolor: str = "white", show_legend: bool = True):
     """
     Saves the visualization of the dotted chart
 
@@ -447,6 +455,8 @@ def save_vis_dotted_chart(log: Union[EventLog, pd.DataFrame], file_path: str, at
     :param log: Event log
     :param file_path: Destination path
     :param attributes: Attributes that should be used to construct the dotted chart (for example, ["concept:name", "org:resource"])
+    :param bgcolor: background color to be used in the dotted chart
+    :param show_legend: boolean (enables/disables showing the legend)
 
     .. code-block:: python3
 
@@ -462,16 +472,23 @@ def save_vis_dotted_chart(log: Union[EventLog, pd.DataFrame], file_path: str, at
 
     format = os.path.splitext(file_path)[1][1:].lower()
     log, attributes = __dotted_attribute_selection(log, attributes)
+
+    parameters = {}
+    parameters["format"] = format
+    parameters["bgcolor"] = bgcolor
+    parameters["show_legend"] = show_legend
+
     from pm4py.visualization.dotted_chart import visualizer as dotted_chart_visualizer
-    gviz = dotted_chart_visualizer.apply(log, attributes, parameters={"format": format, "bgcolor": bgcolor})
+    gviz = dotted_chart_visualizer.apply(log, attributes, parameters=parameters)
     dotted_chart_visualizer.save(gviz, file_path)
 
 
-def view_sna(sna_metric: SNA):
+def view_sna(sna_metric: SNA, variant_str: Optional[str] = None):
     """
     Represents a SNA metric (.html)
 
     :param sna_metric: Values of the metric
+    :param variant_str: variant to be used (default: pyvis)
 
     .. code-block:: python3
 
@@ -480,17 +497,27 @@ def view_sna(sna_metric: SNA):
         metric = pm4py.discover_subcontracting_network(dataframe, resource_key='org:resource', timestamp_key='time:timestamp', case_id_key='case:concept:name')
         pm4py.view_sna(metric)
     """
+    if variant_str is None:
+        if constants.DEFAULT_GVIZ_VIEW == "matplotlib_view":
+            variant_str = "networkx"
+        else:
+            variant_str = "pyvis"
+
     from pm4py.visualization.sna import visualizer as sna_visualizer
-    gviz = sna_visualizer.apply(sna_metric, variant=sna_visualizer.Variants.PYVIS)
-    sna_visualizer.view(gviz, variant=sna_visualizer.Variants.PYVIS)
+    variant = sna_visualizer.Variants.PYVIS
+    if variant_str == "networkx":
+        variant = sna_visualizer.Variants.NETWORKX
+    gviz = sna_visualizer.apply(sna_metric, variant=variant)
+    sna_visualizer.view(gviz, variant=variant)
 
 
-def save_vis_sna(sna_metric: SNA, file_path: str):
+def save_vis_sna(sna_metric: SNA, file_path: str, variant_str: Optional[str] = None):
     """
     Saves the visualization of a SNA metric in a .html file
 
     :param sna_metric: Values of the metric
     :param file_path: Destination path
+    :param variant_str: variant to be used (default: pyvis)
 
     .. code-block:: python3
 
@@ -500,9 +527,20 @@ def save_vis_sna(sna_metric: SNA, file_path: str):
         pm4py.save_vis_sna(metric, 'sna.png')
     """
     file_path = str(file_path)
+
+    if variant_str is None:
+        if constants.DEFAULT_GVIZ_VIEW == "matplotlib_view":
+            variant_str = "networkx"
+        else:
+            variant_str = "pyvis"
+
     from pm4py.visualization.sna import visualizer as sna_visualizer
-    gviz = sna_visualizer.apply(sna_metric, variant=sna_visualizer.Variants.PYVIS)
-    sna_visualizer.save(gviz, file_path, variant=sna_visualizer.Variants.PYVIS)
+    variant = sna_visualizer.Variants.PYVIS
+    if variant_str == "networkx":
+        variant = sna_visualizer.Variants.NETWORKX
+
+    gviz = sna_visualizer.apply(sna_metric, variant=variant)
+    sna_visualizer.save(gviz, file_path, variant=variant)
 
 
 def view_case_duration_graph(log: Union[EventLog, pd.DataFrame], format: str = "png", activity_key="concept:name", timestamp_key="time:timestamp", case_id_key="case:concept:name"):
@@ -851,7 +889,7 @@ def view_ocdfg(ocdfg: Dict[str, Any], annotation: str = "frequency", act_metric:
     :param act_threshold: The threshold to apply on the activities frequency (default: 0). Only activities having a frequency >= than this are kept in the graph.
     :param edge_threshold: The threshold to apply on the edges frequency (default 0). Only edges having a frequency >= than this are kept in the graph.
     :param performance_aggregation: The aggregation measure to use for the performance: mean, median, min, max, sum
-    :param format: The format of the output visualization
+    :param format: The format of the output visualization (if html is provided, GraphvizJS is used to render the visualization in an HTML page)
     :param bgcolor: Background color of the visualization (default: white)
 
     .. code-block:: python3
@@ -923,7 +961,7 @@ def view_ocpn(ocpn: Dict[str, Any], format: str = constants.DEFAULT_FORMAT_GVIZ_
     Visualizes on the screen the object-centric Petri net
 
     :param ocpn: Object-centric Petri net
-    :param format: Format of the visualization
+    :param format: Format of the visualization (if html is provided, GraphvizJS is used to render the visualization in an HTML page)
     :param bgcolor: Background color of the visualization (default: white)
 
     .. code-block:: python3
@@ -968,7 +1006,7 @@ def view_network_analysis(network_analysis: Dict[Tuple[str, str], Dict[str, Any]
 
     :param network_analysis: Network analysis
     :param variant: Variant of the visualization: - frequency (if the discovered network analysis contains the frequency of the interactions) - performance (if the discovered network analysis contains the performance of the interactions)
-    :param format: Format of the visualization
+    :param format: Format of the visualization (if html is provided, GraphvizJS is used to render the visualization in an HTML page)
     :param activity_threshold: The minimum number of occurrences for an activity to be included (default: 1)
     :param edge_threshold: The minimum number of occurrences for an edge to be included (default: 1)
     :param bgcolor: Background color of the visualization (default: white)
@@ -1019,7 +1057,7 @@ def view_transition_system(transition_system: TransitionSystem, format: str = co
     Views a transition system
 
     :param transition_system: Transition system
-    :param format: Format of the visualization
+    :param format: Format of the visualization (if html is provided, GraphvizJS is used to render the visualization in an HTML page)
     :param bgcolor: Background color of the visualization (default: white)
 
     .. code-block:: python3
@@ -1063,7 +1101,7 @@ def view_prefix_tree(trie: Trie, format: str = constants.DEFAULT_FORMAT_GVIZ_VIE
     Views a prefix tree
 
     :param prefix_tree: Prefix tree
-    :param format: Format of the visualization
+    :param format: Format of the visualization (if html is provided, GraphvizJS is used to render the visualization in an HTML page)
     :param bgcolor: Background color of the visualization (default: white)
 
     .. code-block:: python3
@@ -1151,7 +1189,7 @@ def save_vis_alignments(log: Union[EventLog, pd.DataFrame], aligned_traces: List
     visualizer.save(gviz, file_path)
 
 
-def view_footprints(footprints: Union[List[Dict[str, Any]], Dict[str, Any]], format: str = "png"):
+def view_footprints(footprints: Union[Tuple[Dict[str, Any], Dict[str, Any]], Dict[str, Any]], format: str = "png"):
     """
     Views the footprints as a figure
 
@@ -1169,11 +1207,16 @@ def view_footprints(footprints: Union[List[Dict[str, Any]], Dict[str, Any]], for
     format = str(format).lower()
 
     from pm4py.visualization.footprints import visualizer as fps_visualizer
-    gviz = fps_visualizer.apply(footprints, parameters={"format": format})
+
+    if isinstance(footprints, dict):
+        gviz = fps_visualizer.apply(footprints, parameters={"format": format})
+    else:
+        gviz = fps_visualizer.apply(footprints[0], footprints[1], variant=fps_visualizer.Variants.COMPARISON_SYMMETRIC, parameters={"format": format})
+
     fps_visualizer.view(gviz)
 
 
-def save_vis_footprints(footprints: Union[List[Dict[str, Any]], Dict[str, Any]], file_path: str):
+def save_vis_footprints(footprints: Union[Tuple[Dict[str, Any], Dict[str, Any]], Dict[str, Any]], file_path: str):
     """
     Saves the footprints' visualization on disk
 
@@ -1192,7 +1235,12 @@ def save_vis_footprints(footprints: Union[List[Dict[str, Any]], Dict[str, Any]],
     format = os.path.splitext(file_path)[1][1:].lower()
 
     from pm4py.visualization.footprints import visualizer as fps_visualizer
-    gviz = fps_visualizer.apply(footprints, parameters={"format": format})
+
+    if isinstance(footprints, dict):
+        gviz = fps_visualizer.apply(footprints, parameters={"format": format})
+    else:
+        gviz = fps_visualizer.apply(footprints[0], footprints[1], variant=fps_visualizer.Variants.COMPARISON_SYMMETRIC, parameters={"format": format})
+
     fps_visualizer.save(gviz, file_path)
 
 
@@ -1202,7 +1250,7 @@ def view_object_graph(ocel: OCEL, graph: Set[Tuple[str, str]], format: str = con
 
     :param ocel: object-centric event log
     :param graph: object graph
-    :param format: format of the visualization
+    :param format: format of the visualization (if html is provided, GraphvizJS is used to render the visualization in an HTML page)
     :param bgcolor: Background color of the visualization (default: white)
 
     .. code-block:: python3