+++ pm4py/visualization/common/save.py
@@ -16,7 +16,7 @@
 '''
 import shutil
 import os
-from pm4py.visualization.common import dot_util
+from pm4py.visualization.common import dot_util, html
 
 
 def save(gviz, output_file_path, parameters=None):
@@ -33,5 +33,11 @@ def save(gviz, output_file_path, parameters=None):
     format = os.path.splitext(output_file_path)[1][1:].lower()
     is_dot_installed = dot_util.check_dot_installed()
 
-    render = gviz.render(cleanup=True)
-    shutil.copyfile(render, output_file_path)
+    if format.startswith("html"):
+        html.save(gviz, output_file_path, parameters=parameters)
+    else:
+        render = gviz.render(cleanup=True)
+        shutil.copyfile(render, output_file_path)
+    """elif not is_dot_installed:
+        raise Exception("impossible to save formats different from HTML without the Graphviz binary")"""
+