+++ pm4py/visualization/dotted_chart/visualizer.py
@@ -24,6 +24,7 @@ from pm4py.objects.conversion.log import converter as log_converter
 from pm4py.objects.log.obj import EventLog
 from pm4py.util import exec_utils, vis_utils
 from pm4py.visualization.dotted_chart.variants import classic
+from pm4py.util import constants
 
 
 class Variants(Enum):
@@ -71,6 +72,16 @@ def view(figure: str):
     figure
         Path to the dotted chart
     """
+    if constants.DEFAULT_GVIZ_VIEW == "matplotlib_view":
+        import matplotlib.pyplot as plt
+        import matplotlib.image as mpimg
+        img = mpimg.imread(figure)
+        plt.axis('off')
+        plt.tight_layout(pad=0, w_pad=0, h_pad=0)
+        plt.imshow(img)
+        plt.show()
+        return
+
     if vis_utils.check_visualization_inside_jupyter():
         vis_utils.view_image_in_jupyter(figure)
     else: